Make box rendering part of the API.
